omni-agent/pyproject.toml
[project]
name = "omni-agent"
version = "0.1.0"
description = "Full-featured AI agent with multimodal web browsing, file processing, and terminal execution"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.9.0",
    "asyncio-mqtt>=0.16.0",
    "beautifulsoup4>=4.12.0",
    "click>=8.1.0",
    "fastapi>=0.104.0",
    "httpx>=0.25.0",
    "loguru>=0.7.0",
    "opencv-python>=4.8.0",
    "pillow>=10.0.0",
    "playwright>=1.40.0",
    "pydantic>=2.5.0",
    "python-multipart>=0.0.6",
    "requests>=2.31.0",
    "selenium>=4.15.0",
    "uvicorn>=0.24.0",
    "websockets>=12.0",
    "rich>=13.7.0",
    "typer>=0.9.0",
    "anthropic>=0.34.0",
    "python-dotenv>=1.0.0",
    "redis>=5.0.0",
    "psycopg2-binary>=2.9.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.7.0",
]

[project.scripts]
omni-agent = "omni_agent.main:app"

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true